feat: Move tests to corresponding folder
